Refactor 학사일정 로직 학사일정 리듀서로 옮기기
