+++ package-kuring/Sources/Features/NoticeFeatures/NoticeApp.swift
@@ -11,20 +11,19 @@ import LoginFeatures
 import DepartmentFeatures
 import SubscriptionFeatures
 import ComposableArchitecture
+import AcademicCalendarFeatures
 
 @Reducer
 public struct NoticeAppFeature {
     @ObservableState
     public struct State: Equatable {
-        /// 학사일정 바텀시트
-        public var isAcademicSchedulePresented: Bool = false
-        
         // MARK: 네비게이션
         /// 루트
         public var noticeList = NoticeListFeature.State()
         /// 스택 네비게이션
         public var path = StackState<Path.State>()
         public var signup = EmailVerificationFeature.State()
+        public var academicCalendar = AcademicCalendarFeature.State()
         /// 트리 네비게이션 - ``SubscriptionAppFeature``
         @Presents public var changeSubscription: SubscriptionAppFeature.State?
         
@@ -44,38 +43,10 @@ public struct NoticeAppFeature {
                 print("북마크 가져오기를 실패했어요: \(error.localizedDescription)")
             }
         }
-        
-        mutating func fetchAcademicSchedule() -> AcademicScheduleEntity? {
-            @Dependency(\.academicSchedules) var academicDB
-            
-            do {
-                let schedule = try academicDB.fetch(.init())
-                return schedule
-            } catch {
-                print("❌ 캐싱된 학사 일정을 가져오는데 실패 했습니다: \(error)")
-                return nil
-            }
-        }
-        
-        mutating func updateNewSchedule(from apiEvents: [AcademicEvent], isComplete: Bool = false) {
-            @Dependency(\.academicSchedules) var academicDB
-
-            let entities = apiEvents.map(AcademicEventEntity.init(from:))
-            let schedule = AcademicScheduleEntity(lastUpdated: .now, events: entities)
-            schedule.isComplete = isComplete
-            
-            do {
-                try academicDB.add(schedule)
-                print("✅ \(entities.count)개의 학사 일정 추가를 성공했습니다")
-            } catch {
-                print("❌ 힉사일정을 SwiftData에 추가하는데 실패했습니다: \(error)")
-            }
-        }
     }
 
     public enum Action: BindableAction, Equatable {
         case binding(BindingAction<State>)
-        case onAppear
         /// 루트(``NoticeListFeature``) 액션
         case noticeList(NoticeListFeature.Action)
 
@@ -83,6 +54,7 @@ public struct NoticeAppFeature {
         case path(StackAction<Path.State, Path.Action>)
         /// 이메일 인증 네비게이션
         case signup(EmailVerificationFeature.Action)
+        case academicCalendar(AcademicCalendarFeature.Action)
 
         /// 구독 변경 버튼을 탭한 경우
         case changeSubscriptionButtonTapped
@@ -91,39 +63,11 @@ public struct NoticeAppFeature {
         case changeSubscription(PresentationAction<SubscriptionAppFeature.Action>)
         
         case updateBookmarks(_ notice: Notice, _ isBookmarked: Bool)
-        
-        case toggleAcademicScheduleSheet
-        /// 1달치 학사 일정을 가져옵니다
-        case fetch1MonthAcademicSchedule
-        /// 전체 학사 일정을 가져옵니다
-        case fetchEntireAcademicSchedule
-        /// 최신 학사 일정만 가져옵니다
-        case fetchLatestAcademicSchedule
-        case fetchAcademicScheduleResponse(Result<[AcademicEvent], CalendarKuringError>, _ isComplete: Bool)
-        
-        public enum CalendarKuringError: Error, Equatable {
-            case error(String)
-            
-            public static func == (lhs: CalendarKuringError, rhs: CalendarKuringError) -> Bool {
-                switch (lhs, rhs) {
-                case let (.error(lmsg), .error(rmsg)):
-                    return lmsg == rmsg
-                }
-            }
-        }
     }
 
     @Dependency(\.bookmarks) var bookmarks
     @Dependency(\.departments) var departments
-    @Dependency(\.kuringLink) private var kuringLink
-    
-    private var dateFormatter: DateFormatter {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar.current
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter
-    }
-    
+
     public var body: some ReducerOf<Self> {
         Scope(state: \.noticeList, action: \.noticeList) {
             NoticeListFeature()
@@ -133,40 +77,14 @@ public struct NoticeAppFeature {
             EmailVerificationFeature()
         }
         
+        Scope(state: \.academicCalendar, action: \.academicCalendar) {
+            AcademicCalendarFeature()
+        }
+        
         BindingReducer()
         
         Reduce { state, action in
             switch action {
-            case .onAppear:
-                // 캐싱된 학사일정이 없을시,
-                // 1달치 일정을 먼저 가져오고 바텀시트를 노출한 후, 모든 일정을 가져온다
-                guard let schedule = state.fetchAcademicSchedule() else {
-                    return .concatenate([
-                        .send(.fetch1MonthAcademicSchedule),
-                        .send(.fetchEntireAcademicSchedule)
-                    ])
-                }
-                
-                // 캐싱된 학사일정이 있다
-                if !schedule.events.isEmpty {
-                    let sevenDaysInSeconds: TimeInterval = 7 * 24 * 60 * 60
-                    let rightNow = Date().timeIntervalSince1970
-                    let scheduleLastUpdated = schedule.lastUpdated.timeIntervalSince1970
-                    
-                    // 일주일이 지났다
-                    if rightNow - scheduleLastUpdated >= sevenDaysInSeconds {
-                        // 일주일치 새로운 학사일정만 가져온다
-                        return .concatenate([
-                            .send(.fetchLatestAcademicSchedule)
-                        ])
-                    }
-                    // 일주일 안지났으면 아무일도 없음
-                    return .none
-                }
-                return .none
-            case .toggleAcademicScheduleSheet:
-                state.isAcademicSchedulePresented.toggle()
-                return .none
             case .noticeList(.onAppear):
                 // 북마크 상태 동기화
                 @Dependency(\.bookmarks) var bookmarks
@@ -176,51 +94,6 @@ public struct NoticeAppFeature {
                     print("북마크 가져오기를 실패했어요: \(error.localizedDescription)")
                 }
                 return .none
-            case .fetch1MonthAcademicSchedule:
-                return .run { send in
-                    do {
-                        let result = try await kuringLink.fetchAcademicEvents(
-                            dateFormatter.string(from: Date().startDateOfMonth),
-                            dateFormatter.string(from: Date().endDateOfMonth)
-                        )
-                        await send(.fetchAcademicScheduleResponse(.success(result), false))
-                        await send(.toggleAcademicScheduleSheet)
-                    } catch {
-                        await send(.fetchAcademicScheduleResponse(.failure(.error(error.localizedDescription)), false))
-                    }
-                }
-            case .fetchEntireAcademicSchedule:
-                return .run { send in
-                    do {
-                        let result = try await kuringLink.fetchAcademicEvents(nil, nil)
-                        await send(.fetchAcademicScheduleResponse(.success(result), true))
-                    } catch {
-                        await send(.fetchAcademicScheduleResponse(.failure(.error(error.localizedDescription)), false))
-                    }
-                }
-            case .fetchLatestAcademicSchedule:
-                guard let schedule = state.fetchAcademicSchedule() else {
-                    return .none
-                }
-                
-                return .run { send in
-                    do {
-                        let result = try await kuringLink.fetchAcademicEvents(dateFormatter.string(from: schedule.lastUpdated), nil)
-                        await send(.fetchAcademicScheduleResponse(.success(result), true))
-                        await send(.toggleAcademicScheduleSheet)
-                    } catch {
-                        await send(.fetchAcademicScheduleResponse(.failure(.error(error.localizedDescription)), false))
-                    }
-                }
-            case .fetchAcademicScheduleResponse(let result, let isComplete):
-                switch result {
-                case .success(let events):
-                    state.updateNewSchedule(from: events, isComplete: isComplete)
-                    return .none
-                case .failure(let error):
-                    print("Error: \(error)")
-                    return .none
-                }
             case let .path(.element(id: _, action: .detail(.delegate(action)))):
                 switch action {
                 case let .bookmarkUpdated(notice, isBookmarked):
@@ -355,7 +228,7 @@ public struct NoticeAppFeature {
                     )
                 )
                 return .none
-            case .path, .noticeList, .changeSubscription, .signup, .binding:
+            case .path, .noticeList, .changeSubscription, .signup, .binding, .academicCalendar:
                 return .none
             }
         }
